Extending the runtime error raising framework to all actors (#938)
I have changed the AmberException to WorkflowRuntimeException which takes WorkflowRuntimeError as a parameter. Whenever a piece of code ends up in an error, it raises a WorkflowRuntimeException. If the error is raised in an actor other than controller, it send a message to the controller with the WorkflowRuntimeError. When Controller sees such an error either originating within itself, or received through message, it calls eventListener's lambda to send the message to the UI through the socket.