Infer AttributeType when scan CSV file (#1022)
### This PR
Infers the schema (columns' AttributeType) when reading from a CSV file by reading first N(default = 100) records
### AttributeType
By default, when we read from the csv. file, all data types are `AttributeType.String`.
There will be a a Ranking in current AttributeType

         String/Timestamp, Boolean, Double, Long, Integer 

### Logic:
#### Infer Order: 
`Integer->Long->Double->Boolean->String/Timestamp`
we will try to parse the data by Infer Order
For example:
        we will first try to do parse the data to the Integer, if it is success, we will assume the data is Integer
        if it throw out exception, it will continue to try to parse the data to the Long....

#### Priority of the attributeType: 
`String/Timestamp->Boolean->Double->Long->Integer`
Then, when we infer several rows data, we will try to merge them by Priority of the attributeType
For example:
        [String, Integer, Double, Integer]
        [Integer, Integer, Double, Long]
        [Long, Integer, Long, Long]
will become
        [String,Integer,Double,Long]

### Testing
- Currently, there is no way to directly show column's type to view current column's type
- We can use Type Casting operator to view column's type, just connecting them like this


<img width="684" alt="Screen Shot 2021-02-17 at 7 16 59 PM" src="https://user-images.githubusercontent.com/41406160/108300169-b7d03800-7154-11eb-8118-2574fb318aaf.png">
We can see input column's Attribute Type on the right panel.
<img width="253" alt="Screen Shot 2021-02-17 at 7 17 44 PM" src="https://user-images.githubusercontent.com/41406160/108300229-d20a1600-7154-11eb-89a1-faee46baf9ef.png">

### For the issues related to parsing 1, 0 to boolean:
We decide to parse it to boolean
The reasons are below:
Right now, since the AttributeType can only downgrade, I can only consider the way like the diagram below, but this ways may contain some problem that once we find 1 or 0 in a column of data, we can only end up with boolean or string. If the last row is 1 or 0, it will cause the problem.
<img width="531" alt="Screen Shot 2021-02-18 at 11 39 10 AM" src="https://user-images.githubusercontent.com/41406160/108411810-ee05ca00-71dd-11eb-8447-8105314609b7.png">

Other than that, What I am considering is using a special string to record current type，and have a for loop to transfer the string to the AttributeType Array. The picture below shows the process, but this will make the Code becomes strange.
<img width="559" alt="Screen Shot 2021-02-18 at 11 42 06 AM" src="https://user-images.githubusercontent.com/41406160/108412152-5654ab80-71de-11eb-8fc0-ab7dac0895d8.png">

Moreover, the TypeCasting Operator can transfer 0, 1 column to boolean. So, users can easily transfer the 1, 0 column to the BooleanType Easily.
<img width="1400" alt="Screen Shot 2021-02-18 at 11 47 51 AM" src="https://user-images.githubusercontent.com/41406160/108412832-24901480-71df-11eb-802c-30dc916cc524.png">
<img width="1311" alt="Screen Shot 2021-02-18 at 11 48 23 AM" src="https://user-images.githubusercontent.com/41406160/108412902-35d92100-71df-11eb-86b0-586c695b54c5.png">
<img width="1146" alt="Screen Shot 2021-02-18 at 11 48 42 AM" src="https://user-images.githubusercontent.com/41406160/108412942-4093b600-71df-11eb-9c72-e000be3f8fb9.png">



Co-authored-by: Yicong Huang <hyc541978023@gmail.com>