Incrementalize word cloud operator (Step 2) (#961)
This is the part 2 of the Word Cloud incrementalization project. See https://github.com/Texera/texera/issues/936

After incrementalize the word cloud operator, we want to pass the results update from the worker to the controller and web server.

We will reuse the existing logic of QueryStatistics. Whenever the controller query for statistics updates, it will also ask if there are any new updates for the output data if the operator is a sink operator.

Old design:

![Texera Overall Infrastructure-WordCloud](https://user-images.githubusercontent.com/12578068/104516487-4297a300-55a9-11eb-9869-9116dac5e7ed.png)


New design, differences marked in red:
![Texera Overall Infrastructure-WordCloud (1)](https://user-images.githubusercontent.com/12578068/104516500-462b2a00-55a9-11eb-850b-630b614f549d.png)