Add Query Operator and Worker Result Control Message (#1100 Part 5) (#1203)
This is Part 5 of the PRs to break down the big PR #1100, which aims to provide incremental result update in delta mode.


Previously in #1199, we removed the query result functionaility from query statistics. In this PR, we add a new control message `QueryResults` for both controller and worker. The purpose of these control messages are purely to query workflow results and send the result update to the frontend.