Add Python Implementation of DoubleBlockingQueue (#1262)
This PR adds a new Implementation of the DoubleBlockingQueue. This queue is used to process Control Message and Data Messages differently, fitting the need of DP and NetworkSender.

Internally, it has a main queue and a sub queue. We give the following definitions:
- the main queue has a higher priority than the sub queue;
- the sub queue can be disabled. 

Thus, the main queue is suitable for Control Messages, and sub queue is suitable for Data Messages.

Some public APIs:
1. `__init__(*subtypes: tuple[type])` when construct, provide the types that belongs to sub types. 
2. `put(item: T) -> None`: put an item of any type. 
3. `get() -> T`: blocking get, until the next available item, either from the main queue or from the enabled sub queue.
4. `empty() -> bool`: check if the entire queue is empty. Disabled sub queue will not be counted.
5. `main_empty() -> bool`: check if the main queue is empty.
6. `sub_empty() -> bool`: check if the sub queue is empty.
7. `disable_sub() -> None` and `enable_sub() -> None`: to toggle on/off the sub queue, when disabled, sub queue elements will not be emitted.