Refactoring Tuple in Python UDF (#1385)
Currently, the `Python UDF V2` operator converts the input `pyarrow.Table` to `pandas.DataFrame` then to a list of `Tuple`. This becomes the main bottleneck in many of the existing workflows.

This PR:
1. refactors Tuple in `Python UDF V2` by removing the transformation from the `pyarrow.Table` to `pandas.DataFrame`. Also makes the field access and output tuple creation "lazy".
2. Pass the output schema into python UDF through the `InitializeOperatorLogic` control message to enable some optimization by enforcing a fixed schema on output tuples. Currently, we are only passing the schema field names.

Result:
On my local machine, one workflow with 3 python UDF operators took ~200s to complete without this change.
After applying this change, the same workflow took ~ the 40s. 

Co-authored-by: Yicong Huang <17627829+Yicong-Huang@users.noreply.github.com>