Reshape 4/5: Framework to sample workloads (#1410)
This framework adds the functionality to sample workloads. A sender worker records the sample workloads of its receivers in its partitioner class. This only happens if the partitioner shuffles data across nodes. Whenever the controller contacts the worker to get metrics, it gives the entire sample collected till now to the controller. Both the controller and the workers only record a fixed number of samples to not use too much memory.