Add project management feature (#1360)
### Introduction
---
This PR introduces a "user project" feature which acts as a label to help organize a user's workflows and files.  As the number of workflows and files of a user begins to grow too large to manage individually, the user may group them into various, easier-to-manage projects.  The main goals of this PR are to allow a user to:

- Create / rename / delete projects
- View a list of their projects
- Add workflows / files to a project
- Remove workflows / files from a project
- View a particular project and its workflows / files

### Solution
---
The main design goal is to minimize cluttering in the UserDashboard component by having each project display only relevant information, such as only the workflows and files that belong to that project.  Also, rather than displaying all project information in a single component, there are two components that separate seeing a user's list of projects from seeing what a single project contains.

To address these requirements, a new feature has been added to the User Dashboard, and a typical use case is as follows: 
- Navigate to the User Dashboard
- Click on "User Projects" from the Feature Bar
- View / rename / delete / sort the user's list of projects
- Click on a particular project
- View / add / remove workflows and files of that particular project
- Manage the workflows / files themselves from within a project

Currently, projects are backward compatible with workflows and files (workflows / files features function completely independently of projects).  Workflows and files do not have to be associated with a user project.

On the database layer, a new `user_project` table was created to store information about each project: a project's id, a project owner's id, the name of the project, and its time of creation.  New tables `workflow_of_project` and `file_of_project` were also created to store mappings between a project and the workflows / files it contains.  These mapping tables have dependencies on the `user_project`, `workflow`, and `user_file` tables, and any deletions will cascade to the mappings as well.

A new backend API endpoint was created at 'api/project' in `ProjectResource.scala` handling requests from the frontend.  

On the frontend, the following were created to implement the user project feature:
- `UserProjectListComponent`: a child component of `UserDashboard` which lists all the projects of a user
- `UserProjectSectionComponent` : another child component of `UserDashboard` listing a single project and the workflows / files it contains
- `UserProjectService`: handling requests to backend API 
- `NgbdModalAddProjectFileComponent`, `NgbdModalRemoveProjectFileComponent`, `NgbdModalAddProjectWorkflowComponent`, `NgbdModalRemoveProjectWorkflowComponent` : modals to enable adding / removing workflows / files of project

### Demo
---
https://user-images.githubusercontent.com/57075492/151860398-487346e4-0588-410b-88e4-538e024e4228.mov

### Current Issues 
---
Currently, there is some usage of non read-only data structures to enable sorting of projects / workflows, but ideally, it should be fixed to be ReadOnly.  There are also some instances of duplicate logic introduced such as in the `UserProjectSectionComponent` to avoid updating the `SavedWorkflowSectionComponent` and `UserFileSectionComponent`.  However, to improve maintainability such components should be modified to support user projects and reused. 

### Future TODOs
---
The basic features of a user project are included in this PR, but other immediate goals include refactoring the current code to reduce duplicate logic and switching to purely read-only data structures.  Enhancements to the user project feature also include introducing colors to projects to allow for clearer visual indication of organization and potentially introducing sharing of projects.

Co-authored-by: Shengquan Ni <13672781+shengquan-ni@users.noreply.github.com>
Co-authored-by: Yicong Huang <17627829+Yicong-Huang@users.noreply.github.com>